+++ example/lib/game_tiled_map.dart
@@ -1,0 +1,55 @@
+import 'package:bonfire/bonfire.dart';
+import 'package:bonfire/tiled/tiled_world_map.dart';
+import 'package:example/decoration/barrel_dragable.dart';
+import 'package:example/decoration/torch.dart';
+import 'package:example/enemy/goblin.dart';
+import 'package:example/interface/knight_interface.dart';
+import 'package:example/map/dungeon_map.dart';
+import 'package:example/player/knight.dart';
+import 'package:flutter/material.dart';
+
+class GameTiledMap extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return BonfireTiledWidget(
+      joystick: Joystick(
+        directional: JoystickDirectional(
+          spriteBackgroundDirectional: Sprite('joystick_background.png'),
+          spriteKnobDirectional: Sprite('joystick_knob.png'),
+          size: 100,
+          isFixed: false,
+        ),
+        actions: [
+          JoystickAction(
+            actionId: 0,
+            sprite: Sprite('joystick_atack.png'),
+            align: JoystickActionAlign.BOTTOM_RIGHT,
+            size: 80,
+            margin: EdgeInsets.only(bottom: 50, right: 50),
+          ),
+          JoystickAction(
+            actionId: 1,
+            sprite: Sprite('joystick_atack_range.png'),
+            spriteBackgroundDirection: Sprite('joystick_background.png'),
+            enableDirection: true,
+            size: 50,
+            margin: EdgeInsets.only(bottom: 50, right: 160),
+          )
+        ],
+      ),
+      player: Knight(
+        Position((8 * DungeonMap.tileSize), (12 * DungeonMap.tileSize)),
+      ),
+      interface: KnightInterface(),
+      tiledMap: TiledWorldMap(
+        'tiled/mapa1.json',
+        forceTileSize: DungeonMap.tileSize,
+      )
+        ..registerObject('goblin', (x, y) => Goblin(Position(x, y)))
+        ..registerObject('torch', (x, y) => Torch(Position(x, y)))
+        ..registerObject('barrel', (x, y) => BarrelDraggable(Position(x, y))),
+      background: BackgroundColorGame(Colors.blueGrey[900]),
+      lightingColorGame: Colors.black.withOpacity(0.5),
+    );
+  }
+}
